Fix kdeinit module lookup
KLibrary's lookup magic is not so magic these days - is just uses
QCoreApplication::libraryPaths, which is not what we want.  Instead, we
let dlopen() do the searching for us, plus hack in a check in the
library installation directory for kinit (since dlopen() called from Qt
does not respect kdeinit5's RUNPATH).

This should cover most common cases (module installed to standard
location, module installed to same location as the kinit framework,
module in LD_LIBRARY_PATH), and if it still fails we just fall back to
the normal executable.

REVIEW: 116927