Make kdeinit_exec_wait work again (wait until process finishes)
In revision 6eb89d5d12ad1aab28632b8cadfb3ed44d9b9ee6 DBusWait was
removed because nobody was using it from the outside world but we
were using it from within klauncher in the case that wait was true.

Since the calls to kdeinit_exec_wait now use DBusNone we do not wait
until the process dies to finish the request.

This commit adds a new member KLauncherRequest->wait which is set to
true in the case of kdeinit_exec_wait and makes the old code path
work again.

CCMAIL: dfaure@kde.org