Replace GPL proctitle code with BSD-licensed code from OpenSSH"
This reverts commit b15084d9316e34462c2508051c6ac89bcf8d77ee.

It seems to have been a false analysis of the problem; the kernel's view
of the command line (as exposed in /proc/self/environ) is messed up with
both proctitle versions, although the new code makes it more obvious. In
both cases, though, the libc environ variable (which is what code
actually uses, including getenv(), putenv() and exec() calls) is updated
with a copy of the original environment.

CCMAIL: hrvoje.senjan@gmail.com
CCMAIL: afiestas@kde.org