If OOM protection is on, drop privileges in start_kdeinit!
Due to a build system screwup, the correct value of OOM protection
enabled was used to determine if start_kdeinit should be installed
setuid-root, and the WRONG value was used to determine whether
privileges should be dropped, i.e. they were not dropped. This
triggered Qt's setuid abort.
I can't explain how this went undetected for so long, or at all.