Add more debug info, useful for when a start_service does not work
If a start_service gets stuck (requestDone is never called) the
phases will never end, resulting in services not being executed.

The code is a hell and it is impossible to add any kind of fallback for
when a start_service does not work (at least not without doing major
changes to the code).

So at the moment best we can do is add a debug statement that will
help debug cases when startup gets stucked.