kinit: Silence Coverity warning about socket filename.
Coverity CID 258481

Basically, our own name buffer was longer than sockaddr_un had room for.
While this didn't cause bugs in practice, it's probably possible in
theory to somehow come up with a socket path longer than 108 bytes.

And either way, we're already checking for this, the check might as well
be accurate.